chore: move removetest.go to the internal package
* Move the edit/remove_test/removetest.go file to the internal package, as it is
  intended to aid testing.
* Rename the method ExecuteTestCases to ExecuteRemoveTestCases.
