+++ kustomize/commands/edit/remove/removetransformer_test.go
@@ -6,15 +6,15 @@ package remove
 import (
 	"testing"
 
-	"sigs.k8s.io/kustomize/kustomize/v5/commands/edit/remove_test"
+	"sigs.k8s.io/kustomize/kustomize/v5/commands/internal/remove"
 	"sigs.k8s.io/kustomize/kyaml/errors"
 )
 
 func TestRemoveTransformer(t *testing.T) {
-	testCases := []remove_test.Case{
+	testCases := []remove.Case{
 		{
 			Description: "remove transformers",
-			Given: remove_test.Given{
+			Given: remove.Given{
 				Items: []string{
 					"transformer1.yaml",
 					"transformer2.yaml",
@@ -22,7 +22,7 @@ func TestRemoveTransformer(t *testing.T) {
 				},
 				RemoveArgs: []string{"transformer1.yaml"},
 			},
-			Expected: remove_test.Expected{
+			Expected: remove.Expected{
 				Items: []string{
 					"transformer2.yaml",
 					"transformer3.yaml",
@@ -34,7 +34,7 @@ func TestRemoveTransformer(t *testing.T) {
 		},
 		{
 			Description: "remove transformer with pattern",
-			Given: remove_test.Given{
+			Given: remove.Given{
 				Items: []string{
 					"foo/transformer1.yaml",
 					"foo/transformer2.yaml",
@@ -43,7 +43,7 @@ func TestRemoveTransformer(t *testing.T) {
 				},
 				RemoveArgs: []string{"foo/transformer*.yaml"},
 			},
-			Expected: remove_test.Expected{
+			Expected: remove.Expected{
 				Items: []string{
 					"do/not/deleteme/please.yaml",
 				},
@@ -56,7 +56,7 @@ func TestRemoveTransformer(t *testing.T) {
 		},
 		{
 			Description: "nothing found to remove",
-			Given: remove_test.Given{
+			Given: remove.Given{
 				Items: []string{
 					"transformer1.yaml",
 					"transformer2.yaml",
@@ -64,7 +64,7 @@ func TestRemoveTransformer(t *testing.T) {
 				},
 				RemoveArgs: []string{"foo"},
 			},
-			Expected: remove_test.Expected{
+			Expected: remove.Expected{
 				Items: []string{
 					"transformer2.yaml",
 					"transformer3.yaml",
@@ -74,14 +74,14 @@ func TestRemoveTransformer(t *testing.T) {
 		},
 		{
 			Description: "no arguments",
-			Given:       remove_test.Given{},
-			Expected: remove_test.Expected{
+			Given:       remove.Given{},
+			Expected: remove.Expected{
 				Err: errors.Errorf("must specify a transformer file"),
 			},
 		},
 		{
 			Description: "remove with multiple pattern arguments",
-			Given: remove_test.Given{
+			Given: remove.Given{
 				Items: []string{
 					"foo/foo.yaml",
 					"bar/bar.yaml",
@@ -94,7 +94,7 @@ func TestRemoveTransformer(t *testing.T) {
 					"tra*.yaml",
 				},
 			},
-			Expected: remove_test.Expected{
+			Expected: remove.Expected{
 				Items: []string{
 					"do/not/deleteme/please.yaml",
 				},
@@ -107,5 +107,5 @@ func TestRemoveTransformer(t *testing.T) {
 		},
 	}
 
-	remove_test.ExecuteTestCases(t, testCases, "transformers", newCmdRemoveTransformer)
+	remove.ExecuteRemoveTestCases(t, testCases, "transformers", newCmdRemoveTransformer)
 }
